# opt: speed up lookup constraint builder