--- a/pkg/sql/opt/lookupjoin/constraint_builder.go
+++ b/pkg/sql/opt/lookupjoin/constraint_builder.go
@@ -191,7 +191,6 @@ func (b *ConstraintBuilder) Build(
 	// an equality with another column or a constant.
 	numIndexKeyCols := index.LaxKeyColumnCount()
 
-	keyCols := make(opt.ColList, 0, numIndexKeyCols)
 	var derivedEquivCols opt.ColSet
 	// Don't change the selectivity estimate of this join vs. other joins which
 	// don't use derivedFkOnFilters. Add column IDs from these filters to the set
@@ -212,14 +211,16 @@ func (b *ConstraintBuilder) Build(
 		}
 	}
 
-	rightSideCols := make(opt.ColList, 0, numIndexKeyCols)
+	colsAlloc := make(opt.ColList, numIndexKeyCols*2)
+	keyCols := colsAlloc[0:0:numIndexKeyCols]
+	rightSideCols := colsAlloc[numIndexKeyCols : numIndexKeyCols : numIndexKeyCols*2]
 	var inputProjections memo.ProjectionsExpr
 	var lookupExpr memo.FiltersExpr
 	var allLookupFilters memo.FiltersExpr
 	var filterOrdsToExclude intsets.Fast
 	foundLookupCols := false
 	lookupExprRequired := false
-	remainingFilters := make(memo.FiltersExpr, 0, len(onFilters))
+	var remainingFilters memo.FiltersExpr
 
 	// addEqualityColumns adds the given columns as an equality in keyCols if
 	// lookupExprRequired is false. Otherwise, the equality is added as an
@@ -382,6 +383,7 @@ func (b *ConstraintBuilder) Build(
 			allLookupFilters = append(allLookupFilters, allFilters[filterIdx])
 			filterOrdsToExclude.Add(filterIdx)
 			if remaining != nil {
+				remainingFilters = make(memo.FiltersExpr, 0, len(onFilters))
 				remainingFilters = append(remainingFilters, *remaining)
 			}
 		}
@@ -414,6 +416,9 @@ func (b *ConstraintBuilder) Build(
 	// Reduce the remaining filters.
 	for i := range onFilters {
 		if !filterOrdsToExclude.Contains(i) {
+			if remainingFilters == nil {
+				remainingFilters = make(memo.FiltersExpr, 0, len(onFilters))
+			}
 			remainingFilters = append(remainingFilters, onFilters[i])
 		}
 	}