@mgartner 2023-03-21
I think this should be good to backport to v22.2 after some baking time on the master branch.